// SaintMichael/src/app/models/evaluacion.ts
export interface Evaluacion {
        idCurso: any;
		examen: any;
		quiz: any;
		tarea: any;
		proyecto: any;
		cotidiano: any;
}


export interface Evaluacion2 {
    idEvaluacion: number;
    idCurso: number;
    examen: number;
    quiz: number;
    tarea: number;
    proyecto: number;
    cotidiano: number;
}